food-app/backend/foodapp/requirements.txt
dj-database-url==3.0.0
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-otp==1.6.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
pyotp==2.9.0
rest-framework-simplejwt==0.0.2
stripe==12.2.0
python-dotenv
psycopg2-binary>=2.9.9
Pillow
whitenoise
cloudinary
django-cloudinary-storage